feat: let the AI tool know some tools it can use
Copilot is currently only able to read the code in the same file.
Therefore, we should place the class we want to invoke at the top of the file as a comment. This way, Copilot will be able to generate code using the class we commented on the top, even if we only comment the interface of it.